build(keyring-eth-ledger-bridge): add @types/web to provide DOM type definitions
Some files required "DOM" typing definitions.

This was "working fine" before that commit before since `jsdom` was
implictly used by `tsc` during type checking, allowing to `tsc` to have
some "DOM" type definitions.

Now that `jsdom` is not hoisted to the top-level and stays under
`keyring-eth-trezor` workspace might be reason why we did not needed
this in the first place!