ci: update actions, dependabot config, and internal dependencies (#43)
The current Dependabot configuration is not working. This change copies
the configuration from [MetaMask/core](https://github.com/MetaMask/core)
to attempt to fix the issue.

It also updates some actions and adds the `--immutable-cache` flag to
some `yarn install` invocations.

The `@metamask/keyring-api` dependency used by
`@metamask/eth-snap-keyring` is moved to `peerDependencies` and bumped
to 8.1.1.