chore(keyring-api): update peer dependency `@metamask/providers@^17.2.0` -> `^18.1.0` (#76)
Bump `@metamask/providers` requirement to:
- use `@metamask/rpc-errors` min 7.0.0 for consistent error handling
- not have type error from `@metamask/providers` reference to
readable-stream

## Examples

<!--
Are there any examples of this change being used in another repository?

When considering changes to the MetaMask module template, it's strongly
preferred that the change be experimented with in another repository
first. This gives reviewers a better sense of how the change works,
making it less likely the change will need to be reverted or adjusted
later.
-->