build!: build packages with `ts-bridge` (#40)
This updates all packages to build with `ts-bridge` instead of `tsc`.
This has the benefit that packages will be built as CJS modules and ESM
modules at the same time.

I've also modernised `eth-hd-keyring` in this process, as it was
previously a single JS file. For now it's running TypeScript with
`allowJs`, but this file can be properly converted to TypeScript in the
future.

## Breaking changes

- Each package now has an `exports` field, meaning subpath imports
(`import '@metamask/eth-hd-keyring/dist/foo';`) are no longer possible,
unless explicitly stated in the exports field.

---------

Co-authored-by: Charly Chevalier <charly.chevalier@consensys.net>