feat: add `listAccountTransactions` method (#41)
This method was originally part of the Chain API, but we decided to
migrate it to the Keyring API since it's close related to accounts.

Also, it will allow to list the trasactions for a given account ID
instead of address, which make it simpler to get the balances for
Bitcoin accounts, for example.

Closes: https://github.com/MetaMask/accounts-planning/issues/460
Closes: https://github.com/MetaMask/accounts-planning/issues/453
Replaces: https://github.com/MetaMask/keyring-api/pull/322

---------

Co-authored-by: Charly Chevalier <charly.chevalier@consensys.net>
Co-authored-by: Monte Lai <monte.lai@consensys.net>