release: 18.0.0 (#137)
# Description

This is the release candidate for version 18.0.0. See the CHANGELOGs for
more details.

There was a previous attempt for this release that didn't get through
because of the top-level package bump that was missing, see:
- #136
- #135

> [!WARNING]
> The previous version
[17.0.0](https://github.com/MetaMask/accounts/releases/tag/v17.0.0) did
not report some breaking changes properly. This new version is just
meant to re-align the CHANGELOGs and to fix the versioning.