feat!: use `Messenger` instead of `SnapsController` (#152)
Introducing the concept of `Messenger` (messaging system that we use in
our `core` controllers) to break the runtime dependency we have with the
`SnapController`.

We will also use this new `Messenger` later when the Snap keyring will
re-forward some Snap account events to some other controllers.

This is **BREAKING** because of the removal of the
`KeyringSnapControllerClient` and because the `SnapKeyring`'s
constructor now requires a `Messenger` object instead of the
`SnapController`.