fix: bump `@ethersproject/*` from `5.7.0` to `5.8.0` (#243)
Bumping `@ethersproject/*` since the previous version was pinning a
vulnerable `elliptic` version:
-
https://github.com/ethers-io/ethers.js/blob/v5.7.0/packages/signing-key/package.json#L11

The new `5.8.0` has been released to address this issue:
- https://github.com/ethers-io/ethers.js/releases/tag/v5.8.0

I do not expect any functional changes with this version. The only
reason we are forcing the upgrade is to allow to use `elliptic@6.6.1`
(which is no not vulnerable).