fix(keyring-eth-ledger-bridge): trim `0x` prefix for  `signTransaction` (due to latest `@ethereumjs` upgrade) (#253)
This PR will fix the ledger sign transaction issue due to this
ethereumjs upgrade PR https://github.com/MetaMask/accounts/pull/209

and basically the ethereumjs upgrade code change make the transaction
rawTxHex contain `0x` prefix which is not valid for to
@ledgerhq/hw-eth-app clearSignTransaction function. that functions
expect a rawTxHex without `0x` prefix.
 
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes
[#31004](https://github.com/MetaMask/metamask-extension/issues/31004)
* See: #67890
-->


<!--
Are there any examples of this change being used in another repository?

When considering changes to the MetaMask module template, it's strongly
preferred that the change be experimented with in another repository
first. This gives reviewers a better sense of how the change works,
making it less likely the change will need to be reverted or adjusted
later.
-->