chore: update `@trezor/connect-web` (#241)
Similar to:
- https://github.com/MetaMask/accounts/pull/234

But just use a stricter version range to match the one from the
extension, see:
- https://github.com/MetaMask/metamask-extension/pull/30640

We should remove the `syncpack` rule once we agree on moving on to the
next minor (or higher) of this package. For now it has been decided to
only bump the patch version to avoid introducing too many changes for
Trezor support.

Co-authored-by: Daniel Rocha <68558152+danroc@users.noreply.github.com>