feat: use event payload type for `emitSnapKeyringEvent` (#203)
Currently, the `emitSnapKeyringEvent` iss using a generic payload type
(like a `Record<string, Json>`). We were already using `superstruct`
under the hood to check that payload, but the error was then thrown at
runtime.

This new implementation allows to extract the actual payload type
associated with the event to get errors at compile-time.