feat: add `discoverAccounts` keyring method (#266)
This new method will allow a MetaMask clients to discover potential
accounts that have an history on the blockchain (e.g when importing a
new SRP to MetaMask).

This method won't create the accounts directly, it's the client itself
that will decide to create them or not.

Since discovering accounts might requires some "computation", a Snap
could decide to "cache" the information/data he discovered for all those
accounts, so they can be re-used later on once the account is created.

---------

Co-authored-by: Daniel Rocha <68558152+danroc@users.noreply.github.com>