fix(keyring-eth-ledger-bridge): fix use of `salt` value as a `string` for `signTypedData<V4>` (#249)
This PR is to fix the extension bug
[30473](https://github.com/MetaMask/metamask-extension/issues/30473)

Basically the issue was caused by domain salt was defined as string in
client DApp, and @metamask/eth-sig-util library we used define salt as
arrayType. which cause miss match with previous code of salt value.

This code has been tested with client dapp and our e2e test-dapp to
confirm it will work for both.

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes
[#30473](https://github.com/MetaMask/metamask-extension/issues/30473)
* See: #67890
-->

## Examples

<!--
Are there any examples of this change being used in another repository?

When considering changes to the MetaMask module template, it's strongly
preferred that the change be experimented with in another repository
first. This gives reviewers a better sense of how the change works,
making it less likely the change will need to be reverted or adjusted
later.
-->