refactor: remove timeout for messages requiring user actions (#279)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->

https://github.com/MetaMask/accounts/pull/271 introduced timeouts for
all messages between the Ledger bridge and the iframe. Though a
20-second timeout may not be enough for the user to grab the device and
accept the request - especially if the user wants to look closely at the
signature request on the hardware device.

For this reason, the timeout is being removed for messages requiring
user intervention, while keeping it for messages that do not expect user
actions (i.e. `updateTransportMethod` and `attemptMakeApp`)