feat(keyring-eth-trezor): bump `@trezor/connect-web` to 9.6.0 to support new flow (#300)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->
This PR will upgrade `trezor/connect-web` library based on trezor team
suggestion. please refer to plan
https://github.com/MetaMask/accounts-planning/issues/936 for details.

This PR will add a appName parameter to manifest to initialise
TrezorSDK.

## Examples

<!--
Are there any examples of this change being used in another repository?

When considering changes to the MetaMask module template, it's strongly
preferred that the change be experimented with in another repository
first. This gives reviewers a better sense of how the change works,
making it less likely the change will need to be reverted or adjusted
later.
-->