feat: add qr keyring package (#60)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->

This is a new Keyring implementation compatible with QR-based wallets,
meant to replace `@keystonehq/metamask-airgapped-keyring`.

The goals are:
  - Internalize the package
  - Simplify internal logic (e.g. no event-based logic)
- Remove the need of KeyringController special methods (See methods
nuked in https://github.com/MetaMask/core/pull/6031)
- Provide a consumer-injectable transport bridge to allow clients to
control QR scanning implementation
  - Improve API ergonomics
  - Allow clients to easily support multiple QR devices per user

This package can be tested on Extension through this PR:
https://github.com/MetaMask/metamask-extension/pull/33893

* Fixes https://github.com/MetaMask/accounts/issues/144

## How does it work?

The QrKeyring in this package is split into two main components:
- `QrKeyring`: The main keyring class that exposes the Keyring API. Some
of its logic is implemented in internal utility classes:
- `Device`: A stateful internal wrapper for the keyring's public keys
derivation logic and account management. In the code it's implemented by
a separate class, but it can be seen as an internal component of the
keyring, and It is not exported from the package.
- `QrKeyringBridge`: The interface that a transport bridge must
implement: it allows the client to control how QR codes are scanned.
It's supposed to be implemented by a class that is injected by the
client: when using KeyringController, it can be injected through a
custom keyring builder (See [this keyring
builder](https://github.com/MetaMask/metamask-extension/blob/75576f2d31432563e34fbaa66c7783b77aa5124d/app/scripts/lib/qr-keyring-builder-factory.ts)
that will be used by Extension).
- `QrKeyringScannerBridge`: A generic bridge that allows the client to
pass function hooks at construction time, which will be called when the
keyring needs to scan a QR code.

### Device pairing flow

Pairing is the process of initializing the keyring with a device, which
is done by submitting the device details to the keyring in the form of a
serialized UR, which usually comes from the "sync" QR scanned on the
device. The client can trigger this process in three ways:
- At construction time, by passing a `SerializedUR` object to the
`QrKeyring` constructor.
- In this case, the scan is done before the keyring is constructed, and
thus the keyring will directly receive the QR contents.
- By calling the `QrKeyring.pairDevice()` method, which allows the
client to pass a `SerializedUR` object at any time.
- Same as above, the scan is done outside the keyring, and the keyring
will directly receive the QR contents.
- By calling the `QrKeyring.getFirstPage()` method, which will trigger a
scan request if no device is paired yet.
- In this case, the keyring will request a scan through the bridge, and
the client will implement the scan logic.

#### Extension uses `getFirstPage()` to trigger the pairing flow, which
will then request a scan through the bridge

Like for other Hardare wallets, the pairing flow is initiated in
[`connectHardware()`](https://github.com/MetaMask/metamask-extension/blob/75576f2d31432563e34fbaa66c7783b77aa5124d/app/scripts/metamask-controller.js#L5191)
method. See the sequence diagram below for the pairing flow on
Extension, and see [the Extension `requestScan`
callback](https://github.com/MetaMask/metamask-extension/blob/75576f2d31432563e34fbaa66c7783b77aa5124d/app/scripts/metamask-controller.js#L1112)
for the implementation of the scan logic.

```mermaid
sequenceDiagram
        autonumber
        participant A as MM Extension
        create participant B as QrKeyring
        A->>B: getFirstPage()
        alt if not initialized
        create participant C as Bridge
        B->>C: requestScan(PAIRING)
        C-->>+A: 
        note over A,C: The client controls the bridge and implements the scan logic
        A-->>-C: 
        destroy C
        C->>B: return CBOR 
        create participant D as Int. derivation logic
        B->>D: init(CBOR)
        end
        B->>D: accountAtIndex(0..4)
        destroy D
        D->>B: return accounts
        destroy B
        B->>A: return accounts
```

#### Mobile

TBD - Likely using `.pairDevice()` given the current implementation of
QR interactions on mobile.

### Signing flow

```mermaid
sequenceDiagram
        autonumber
        participant A as Client
        create participant B as QrKeyring
        A->>B: sign{Transaction, PersonalMessage, TypedData}()
        alt not initialized
        B->>A: Fail
        end
        create participant C as Bridge
        B->>C: requestScan(SIGN)
        C-->>+A: 
        note over A,C: The client controls the bridge and implements the scan logic
        A-->>-C: 
        destroy C
        C->>B: return signature CBOR 
        destroy B
        B->>A: return hex signature
```

### Account addition flow
No scan required after the initial pairing.

```mermaid
sequenceDiagram
        autonumber
        participant A as Client
        alt non sequential account addition 
        create participant B as QrKeyring
        A->>B: setAccountToUnlock( i )
        end
        A->>B: addAccounts( n )
        loop n-times
        create participant D as Int. derivation logic
        B->>D: accountAtIndex( i )
        destroy D
        D->>B: return account
        end
        B->>A: return created accounts
```


## Examples

<!--
Are there any examples of this change being used in another repository?

When considering changes to the MetaMask module template, it's strongly
preferred that the change be experimented with in another repository
first. This gives reviewers a better sense of how the change works,
making it less likely the change will need to be reverted or adjusted
later.
-->