feat: revert changes to `AccountProvider` and add `Bip44AccountProvider` type alias (#361)
What is the current state of things and why does it need to change? 
1. It was determined that `AccountProviderType` doesn't work well with
the concept of a custom provider. We wanted to let the
`MultichainAccountWallet` orchestrate the discovery so it was determined
it would be better to keep the function signature as is for
`discoverAndCreateAccounts`.
2. Writing `AccountProvider<Bip44Account<KeyringAccount>>` is a
mouthful.
 
What is the solution your changes offer and how does it work? 
1. Remove the `AccountProviderType` enum, revert the changes made to
`AccountProvider`.
2. Write a type alias: `Bip44AccountProvider`.